fix(data, session, build): 修复保存路径栈/缓冲溢出与 zip 超限截断，新增 -Wconversion 告警闸门
问题修复：
1. 修复保存 .pxc/.pxl 时通道名导致的栈溢出：StoreSession::meta_gen() 的 char meta[300] 被用户可编辑通道名溢出写栈，39 处 sprintf 改为 snprintf，SignalModel 新增 kMaxNameLength=128 并在 set_name() 中截断（截断须先于变更守卫的相等比较），另新增 sanitize_meta_value() 去除 CR/LF，避免通道名向 .pxc 头部注入伪造键值行
2. 修复 SpectrumTrace::format_freq() 精度较大时溢出 char[50]：改用 snprintf 并将 strcat 换成带剩余空间的 strncat；TriggerDock 的 sprintf 一并转 snprintf
3. 修复 ZipMaker 对超过 4 GiB 条目的静默截断：AddFromBuffer() 参数 unsigned int 改 size_t 并显式保护（超限返回 false 并写 m_error），读取侧 unzReadCurrentFile()/ZipInnerFileData 同步处理

构建与工程：
4. 新增严格告警闸门：ENABLE_STRICT_WARNINGS（默认 ON，仅启用 -Wconversion、永不接 -Werror）按目标施加于 pxview-core/pxview-data/pxview-utility/pxview-config（四个目标基线已清零），而非全局 CMAKE_CXX_FLAGS（全项目 9043 条）；ENABLE_OLD_STYLE_CAST_WARNINGS 默认 OFF（glib gmacros.h 噪声且 -isystem 压不住）
5. 依赖头改为 SYSTEM include，第三方诊断不进入本项目告警输出
6. 清理核心层 180 条 -Wconversion 告警（均为类型提升产物而非缺陷），View 层清理 dsosignal_paint.cpp 81 条与 trace.h 25 条

测试与 CI：
7. 新增通道名截断回归测试 SetNameClampsOverlongName（超长截断、截断不破坏 set_name() 变更守卫、正好等于上限原样通过）

+++ PXView/pv/api/mcp_transport.cpp
@@ -105,7 +105,7 @@ bool McpTransport::start()
     connect(_server, &QTcpServer::newConnection,
             this, &McpTransport::on_new_connection);
 
-    if (!_server->listen(QHostAddress::LocalHost, _port)) {
+    if (!_server->listen(QHostAddress::LocalHost, static_cast<quint16>(_port))) {
         delete _server;
         _server = nullptr;
         return false;
@@ -265,7 +265,7 @@ void McpTransport::try_handle_request(QTcpSocket* socket)
     QByteArray data = socket->readAll();
 
     // Find header/body boundary
-    int header_end = data.indexOf("\r\n\r\n");
+    int header_end = static_cast<int>(data.indexOf("\r\n\r\n"));
     if (header_end < 0) {
         // Incomplete headers — wait for more data
         _pending_sockets.insert(socket);
@@ -283,7 +283,7 @@ void McpTransport::try_handle_request(QTcpSocket* socket)
             socket->setProperty("_http_buffer", accumulated);
 
             // Check if we have complete headers now
-            int he = accumulated.indexOf("\r\n\r\n");
+            int he = static_cast<int>(accumulated.indexOf("\r\n\r\n"));
             if (he < 0) return; // Still incomplete
 
             // Check Content-Length
@@ -335,7 +335,7 @@ void McpTransport::try_handle_request(QTcpSocket* socket)
             accumulated.append(socket->readAll());
             socket->setProperty("_http_buffer", accumulated);
 
-            int he = accumulated.indexOf("\r\n\r\n");
+            int he = static_cast<int>(accumulated.indexOf("\r\n\r\n"));
             if (he < 0) return;
 
             QByteArray body = accumulated.mid(he + 4);
@@ -359,7 +359,7 @@ void McpTransport::try_handle_request(QTcpSocket* socket)
 void McpTransport::handle_http_request(QTcpSocket* socket, const QByteArray& data)
 {
     // Find header/body boundary
-    int header_end = data.indexOf("\r\n\r\n");
+    int header_end = static_cast<int>(data.indexOf("\r\n\r\n"));
     if (header_end < 0) {
         send_http_response(socket, 400,
             "{\"jsonrpc\":\"2.0\",\"error\":{\"code\":-32700,\"message\":\"Invalid HTTP request\"},\"id\":null}");
